adjust property panel style
